Fixed Gravity Algebra for AIR
Apparently AIR has a bug where the hitTest method will not work when the
stage is magnified. This demo required a significant rewrite to work
around this bug. Now instead of rolling off the selection to make the
choice menu dissappear, users will need to click on an inert
(background) location. This is an improvement, UI-wise.