Fixed Blackbody Curves and Filters Explorer for AIR
This required a hard-coded stage size. There was also a problem with the
x-axis labels: the first label attached programmatically was not getting
its label properly assigned. I think this may be the same bug as
affected the Heliacal Rising and Big Dipper simulators.

Publishing for Flash Player 10 would fix this bug (this is what I did
for the other sims), but it would break the comboboxes. I didn't notice
that this happened in the earlier fixes.

The fix ultimately applied was to re-assign the label text after
attaching.