Fixed Filters Simulator for AIR
This involved hard-coding a stage size.

I discovered another AIR bug: the comboboxes don't work when the
simulator is zoomed-in in the AIR app. The problem appears to be the
hitTest calls being made in the combobox (the same issue that affected
gravAlgebra). The fix was to write a custom hitTest replacement. This
fix will need to be applied to all files using the Flash MX combobox.