Fixed bugs in Rotating Sky Explorer (celHorComp)...
The constellation drop-down box didn't work in AIR when scaled (due to
hitTest).

Also fixed: it was possible to click the buttons behind the drop down if
the space between the constellation options was clicked.

Not fixed: it is possible to change the location of a constellation star
using the text fields (this is a long-standing bug -- not due to AIR).