Fixed combobox problem in Big Dipper Clock demo...
Although previously updated, the combobox did not have the correct code
update. In order to get the combobox event to fire correctly the
targeted Flash Player version needed to be returned to 6 from 10 (which
it was set in 005-B). This then meant that the label on the timeline
needed to be fixed.